\name{Dermet-package}
\alias{Dermet-package}
\alias{Dermet}
\docType{package}
\title{
\packageTitle{Dermet}
}
\description{
\packageDescription{Dermet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Dermet}
\packageIndices{Dermet}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Dermet}

Maintainer: \packageMaintainer{Dermet}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
